requirements.txt
python-dotenv
openai
yaml
